+++ app/controllers/detail/device.js
@@ -18,9 +18,4 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
-    deviceName: null,
-    guards: null,
-    init: function() {
-        this._super(...arguments);
-    },
 });