+++ tests/integration/components/presence-detail-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('presence-detail', 'Integration | Component | presence detail', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });"
+
+  this.render(hbs`{{presence-detail}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:"
+  this.render(hbs`
+    {{#presence-detail}}
+      template block text
+    {{/presence-detail}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});