+++ doc/changelog.md
@@ -6,5 +6,6 @@
  - Implemented _General_ view with following components:
    - List of all devices. Devices in error are colored. Each device has summary about their guards and alarms.
  - Implemented _Device detail_ view with following components:
+   - Presence information.
    - List all guarded topics.
    - Each guarded topic contains table with status of all topic alarms.